+++ packages/core/src/collection/index.ts
@@ -1489,32 +1489,6 @@ export class Collection<
   }
 }
 
-/**
- * Returns a newly created Collection.
- *
- * A Collection manages a reactive set of Information
- * that we need to remember globally at a later point in time.
- * While providing a toolkit to use and mutate this set of Information.
- *
- * It is designed for arrays of data objects following the same pattern.
- *
- * Each of these data object must have a unique `primaryKey` to be correctly identified later.
- *
- * You can create as many global Collections as you need.
- *
- * [Learn more..](https://agile-ts.org/docs/core/agile-instance/methods#createcollection)
- *
- * @public
- * @param config - Configuration object
- * @param agileInstance - Instance of Agile the Collection belongs to.
- */
-export function createCollection<DataType extends Object = DefaultItem>(
-  config?: CollectionConfig<DataType>,
-  agileInstance: Agile = shared
-): Collection<DataType> {
-  return new Collection<DataType>(agileInstance, config);
-}
-
 export type DefaultItem = Record<string, any>; // same as { [key: string]: any };
 export type CollectionKey = string | number;
 export type ItemKey = string | number;
