+++ packages/core/tests/unit/agile.test.ts
@@ -167,115 +167,6 @@ describe('Agile Tests', () => {
       jest.clearAllMocks(); // Because creating the Agile Instance calls some mocks
     });
 
-    describe('createStorage function tests', () => {
-      beforeEach(() => {
-        jest.spyOn(Shared, 'createStorage');
-      });
-
-      it('should call createStorage', () => {
-        const storageConfig = {
-          prefix: 'test',
-          methods: {
-            get: () => {
-              /* empty function */
-            },
-            set: () => {
-              /* empty function */
-            },
-            remove: () => {
-              /* empty function */
-            },
-          },
-          key: 'myTestStorage',
-        };
-
-        const response = agile.createStorage(storageConfig);
-
-        expect(response).toBeInstanceOf(Storage);
-        expect(Shared.createStorage).toHaveBeenCalledWith(storageConfig);
-      });
-    });
-
-    describe('createState function tests', () => {
-      beforeEach(() => {
-        jest.spyOn(Shared, 'createState');
-      });
-
-      it('should call createState with the Agile Instance it was called on', () => {
-        const response = agile.createState('jeff', { key: 'jeffState' });
-
-        expect(response).toBeInstanceOf(State);
-        expect(Shared.createState).toHaveBeenCalledWith('jeff', {
-          key: 'jeffState',
-          agileInstance: agile,
-        });
-      });
-    });
-
-    describe('createCollection function tests', () => {
-      beforeEach(() => {
-        jest.spyOn(Shared, 'createCollection');
-      });
-
-      it('should call createCollection with the Agile Instance it was called on', () => {
-        const collectionConfig = {
-          selectors: ['test', 'test1'],
-          groups: ['test2', 'test10'],
-          defaultGroupKey: 'frank',
-          key: 'myCoolCollection',
-        };
-
-        const response = agile.createCollection(collectionConfig);
-
-        expect(response).toBeInstanceOf(Collection);
-        expect(Shared.createCollection).toHaveBeenCalledWith(
-          collectionConfig,
-          agile
-        );
-      });
-    });
-
-    describe('createComputed function tests', () => {
-      const computedFunction = () => {
-        // empty
-      };
-
-      beforeEach(() => {
-        jest.spyOn(Shared, 'createComputed');
-      });
-
-      it('should call createComputed with the Agile Instance it was called on (default config)', () => {
-        const response = agile.createComputed(computedFunction, [
-          'dummyDep' as any,
-        ]);
-
-        expect(response).toBeInstanceOf(Computed);
-        expect(Shared.createComputed).toHaveBeenCalledWith(computedFunction, {
-          computedDeps: ['dummyDep' as any],
-          agileInstance: agile,
-        });
-      });
-
-      it('should call createComputed with the Agile Instance it was called on (specific config)', () => {
-        const computedConfig = {
-          key: 'jeff',
-          isPlaceholder: false,
-          computedDeps: ['dummyDep' as any],
-          autodetect: true,
-        };
-
-        const response = agile.createComputed(computedFunction, computedConfig);
-
-        expect(response).toBeInstanceOf(Computed);
-        expect(Shared.createComputed).toHaveBeenCalledWith(computedFunction, {
-          ...computedConfig,
-          ...{
-            agileInstance: agile,
-          },
-        });
-      });
-    });
-
     describe('integrate function tests', () => {
       it('should integrate provided Framework', () => {
         const returnedAgile = agile.integrate(testIntegration);
